move HtmlAnchor rack middleware to tdiary/rack
